% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulR.R
\name{qtable}
\alias{qtable}
\title{qtable}
\usage{
qtable(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{vars}{The variables to aggregate.}

\item{groups}{Variables to group by.}

\item{weight}{A variable to weight results by. This is only applied to the margin.}

\item{margin}{Set to TRUE to generate a margin for the first group.}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\value{
A \code{data.frame} or \code{data.table} with the additional class \code{qtable}.
}
\description{
Generating "quick" tables for one or more variables (does not support mixed types).
By default \code{qtable} generates proportions for \code{factor} and \code{character} vectors,
means for \code{numeric} and \code{integer}, and min/max for \code{Date} (including POSIX)
vectors. It always includes the number of observations for each variable.
When producing wide tables, group counts are separate with \code{/}.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

