% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtable.R
\name{qtable_}
\alias{qtable_}
\title{qtable}
\usage{
qtable_(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{vars}{The variables to aggregate, as a character vector.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\value{
A \code{data.frame} or \code{data.table} with the additional class \code{qtable}.
}
\description{
Generating "quick" tables for one or more variables (does not support mixed types).
By default \code{qtable} generates proportions for \code{factor} and \code{character} vectors,
means for \code{numeric} and \code{integer}, and min/max for \code{Date} (including POSIX)
vectors. It always includes the number of observations for each variable.
When producing wide tables, group counts are separate with \code{/}.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

