% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulR.R
\docType{package}
\name{tabulR}
\alias{tabulR}
\alias{tabulR-package}
\title{tabulR/qtable: Create "quick" tables.}
\description{
This is a tiny package that uses \code{\link[data.table]{data.table}} to generate
a quick margin table or plot for one or more variables, by one or more groups.
See \code{\link{qtable}}, \code{\link{bar_chart}} and \code{\link{line_chart}} for more details.
}
\author{
Kristian D. Olsen
}

