% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtable.R
\name{qtable}
\alias{qtable}
\title{qtable}
\usage{
qtable(df, ..., groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{...}{Unquoted variable names passed to \code{\link[dplyr]{select}}.}
}
\value{
Same as \code{\link{qtable}}.
}
\description{
Same as \code{\link{qtable_}}, except designed for use with \code{dplyr}. It takes
unquoted variable names and supports all dplyr select statements, except renaming in
the select statement itself. \code{qtable} also looks for \code{dplyr} groups if
they are not specified.
}
\details{
Note that \code{groups} and \code{weight} still have to be character vectors.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

